This PR contains two commits cherry-picked from #66924.

From #66924 discussions, it is currently blocked because commit 37c3719
uses tools.ormcache and thus awaits al's validation. 

There are currently two tickets in the support perf pipe that are about
slow Accounting Dashboard opening. According to a local benchmark
(that can be found below), merging these two commits is sufficient
to have a good enough speed-up that would allow to close the
two aforementioned tickets.

Moreover these two commits can be safely deployed to stable.

#### speedup

Customer DB, 65 cash/bank account_journal, 30k account_bank_statement_line,
330k account_move.

- account_journal._get_last_bank_statement (state = "posted") avg time: `119 ms -> 7 ms`.
- account_journal_dashboard._kanban_dashboard (33 journals) avg time: `7.5s -> 1.06s`.
- Accounting App search_read: `12.20s -> 3.40s`.

--
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr