@robodoo 2022-01-13
[Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/82707).